feat: extract tray menu, settings, and GitHub glue from __main__
Hardening item 6, extraction 3 (architecture spec A2). TrayMenu owns
the full right-click menu build, every settings setter, the option
constants, and the output-folder/error dialogs; GitHubTray owns the PR
poller lifecycle, change toasts, menu section, and gh-CLI merge (named
github_tray because github_status.py is the poller itself). The flash
gate moves to icons.FlashController per the extraction-1 deviation
note. __main__.py is now 799 lines - under the ~800 item 6 target with
the lifecycle extraction still to come.

pystray and capture import lazily so both components are testable on
the dependency-light system python. 11 new tests; the load-bearing
one builds the FULL menu against the real default config with
pystray/capture stubbed (menu-build crashes were the 2026-05-07 crash
class) - it caught three transform misses and a stranded _updating
attribute before anything shipped.

System suite 251 passed; pyflakes clean on all extracted modules.
