+++ CLAUDE.md
@@ -24,6 +24,8 @@ When a user reports that a meeting "crashed" or "failed" or "didn't finish", do 
 | `dictation_flow.py` | Dictation workflow component: normal/feature-suggest/overlay dictation, auto-stop cap, discard, crash recovery, history. Feature routing state lives in AppState.feature_suggest. |
 | `meeting_dialogs.py` | Meeting Tk dialogs (name, speaker confirmation, recovery, LLM-unavailable) + shared styling helpers and the ABORT sentinel. All run on the DialogDispatcher thread. |
 | `meeting_flow.py` | Meeting workflow component: record/save/post-process pipeline (MeetingJob steps), crash recovery, re-run speaker ID, rename suggestions, minutes via Claude CLI. |
+| `tray_menu.py` | Tray menu construction + all settings setters + option constants. menu_callback() binds pystray item callbacks. |
+| `github_tray.py` | GitHub PR status tray glue: poller lifecycle, PR toasts, menu section, gh merge. |
 | `state_manager.py` | Observable state machine. AppState dataclass, typed StateEvent, event constants, thread-safe emit(), listener subscriptions (on/on_any), event log ringbuffer. All icon/toast updates flow through here. |
 | `transcribe.py` | WhisperX with persistent model cache. Fast path (dictation) and staged pipeline (meeting) |
 | `worker.py` | Multiprocessing transcription worker. CUDA isolation via spawn context |
