+++ docs/TECHNICAL.md
@@ -14,6 +14,8 @@ Everything below is for developers modifying this app or for an AI assistant (li
 | `dictation_flow.py` | **Dictation workflow.** Normal, feature-suggest, and overlay dictation; auto-stop cap, discard, crash recovery, recent history. Feature routing lives in `AppState.feature_suggest`. | `DictationFlow` class --`toggle()`, `toggle_feature_suggest()`, `discard()`, `recover_dictation()`, `recover_feature()`, `recent_history()` |
 | `meeting_dialogs.py` | **Meeting dialog builders.** Tk dialogs (meeting name, speaker confirmation, recovery naming, LLM-unavailable) run on the DialogDispatcher; shared styling helpers + ABORT sentinel. | `MeetingDialogs` class --`ask_meeting_name()`, `ask_speaker_confirmation()`, `ask_recovery_name()`, `show_llm_unavailable()`; module helpers `sanitize_name()`, `flat_button()` |
 | `meeting_flow.py` | **Meeting workflow.** Toggle/start/stop, save-and-enqueue, sequential post-processing worker driving MeetingJob steps, crash recovery, rename suggestions, minutes generation. | `MeetingFlow` class --`toggle()`, `recover_meetings()`, `recover_meeting_speakers()`, `generate_minutes()`, `start_post_worker()`, `pipeline_idle()` |
+| `tray_menu.py` | **Tray menu + settings.** Full right-click menu build, every settings setter, output-folder dialogs, error popup. Menus swap via MenuRefresher under the tray lock. | `TrayMenu` class --`build()`, `_set_*()`, `_change_output_folder()`; module `menu_callback()` + option constants |
+| `github_tray.py` | **GitHub PR tray glue.** Poller lifecycle, PR change toasts with action buttons, GitHub menu section, gh-CLI merge. | `GitHubTray` class --`start()`, `stop()`, `menu_items()`, `_merge_pr()` |
 | `capture.py` | **Audio recording.** Multi-stream mic + speaker loopback via WASAPI. | `AudioRecorder` class --`start()`, `stop()`, `_mic_callback()`, `_speaker_callback()`, `start_streaming()`, `stop_streaming()` |
 | `transcribe.py` | **WhisperX engine.** Model loading, transcription, alignment, diarization. Two paths: fast (dictation) and full (meeting). | `transcribe_fast(audio_np, model)` -- in-memory numpy to text; `transcribe(audio_path, diarize, model)` -- file-based full pipeline; `_load_model()`, `_load_align_model()` |
 | `worker.py` | **Subprocess entry point.** Runs transcription in isolated process (crash safety). Receives requests via Queue, returns results. | `worker_main()` -- event loop handling `transcribe_fast`, `transcribe`, `reload_model`, `shutdown` requests |
@@ -56,6 +58,8 @@ __main__.py (entry point, UI, hotkeys)
 ├── dictation_flow.py (dictation workflow component)
 ├── meeting_dialogs.py (meeting Tk dialogs)
 ├── meeting_flow.py (meeting workflow component)
+├── tray_menu.py (menu + settings component)
+├── github_tray.py (GitHub PR status glue)
 ├── config.py (settings)
 ├── paths.py (directories)
 ├── logger.py (logging)
